+++ lib/dto/Configuration.ts
@@ -1,0 +1,12 @@
+export interface BodyCustomPropertyDto {
+  propertyKey: string;
+  type: string;
+  propertyValue: string,
+  hashTags: [],
+  label: string,
+  propertyLevel: string
+}
+
+export interface CustomPropertyDto {
+  propertyKey: string;
+}